moved KeyboardEventHandlers to common
